requirements.txt
pyyaml
spotipy
tabulate
tk
